+++ tests/unit/async_/io/test_direct.py
@@ -93,9 +93,11 @@ class AsyncFakeBoltPool(AsyncIOPool):
         self.address = address
 
     async def acquire(
-        self, access_mode=None, timeout=None, database=None, bookmarks=None
+        self, access_mode=None, timeout=None, database=None, bookmarks=None,
+        lifeness_check_timeout=None
     ):
-        return await self._acquire(self.address, timeout)
+        return await self._acquire(self.address, timeout,
+                                   lifeness_check_timeout)
 
 
 @mark_async_test
@@ -144,7 +146,7 @@ def assert_pool_size( address, expected_active, expected_inactive, pool):
 @mark_async_test
 async def test_pool_can_acquire(pool):
     address = ("127.0.0.1", 7687)
-    connection = await pool._acquire(address, timeout=3)
+    connection = await pool._acquire(address, 3, None)
     assert connection.address == address
     assert_pool_size(address, 1, 0, pool)
 
@@ -152,8 +154,8 @@ async def test_pool_can_acquire(pool):
 @mark_async_test
 async def test_pool_can_acquire_twice(pool):
     address = ("127.0.0.1", 7687)
-    connection_1 = await pool._acquire(address, timeout=3)
-    connection_2 = await pool._acquire(address, timeout=3)
+    connection_1 = await pool._acquire(address, 3, None)
+    connection_2 = await pool._acquire(address, 3, None)
     assert connection_1.address == address
     assert connection_2.address == address
     assert connection_1 is not connection_2
@@ -164,8 +166,8 @@ async def test_pool_can_acquire_twice(pool):
 async def test_pool_can_acquire_two_addresses(pool):
     address_1 = ("127.0.0.1", 7687)
     address_2 = ("127.0.0.1", 7474)
-    connection_1 = await pool._acquire(address_1, timeout=3)
-    connection_2 = await pool._acquire(address_2, timeout=3)
+    connection_1 = await pool._acquire(address_1, 3, None)
+    connection_2 = await pool._acquire(address_2, 3, None)
     assert connection_1.address == address_1
     assert connection_2.address == address_2
     assert_pool_size(address_1, 1, 0, pool)
@@ -175,7 +177,7 @@ async def test_pool_can_acquire_two_addresses(pool):
 @mark_async_test
 async def test_pool_can_acquire_and_release(pool):
     address = ("127.0.0.1", 7687)
-    connection = await pool._acquire(address, timeout=3)
+    connection = await pool._acquire(address, 3, None)
     assert_pool_size(address, 1, 0, pool)
     await pool.release(connection)
     assert_pool_size(address, 0, 1, pool)
@@ -184,7 +186,7 @@ async def test_pool_can_acquire_and_release(pool):
 @mark_async_test
 async def test_pool_releasing_twice(pool):
     address = ("127.0.0.1", 7687)
-    connection = await pool._acquire(address, timeout=3)
+    connection = await pool._acquire(address, 3, None)
     await pool.release(connection)
     assert_pool_size(address, 0, 1, pool)
     await pool.release(connection)
@@ -195,7 +197,7 @@ async def test_pool_releasing_twice(pool):
 async def test_pool_in_use_count(pool):
     address = ("127.0.0.1", 7687)
     assert pool.in_use_connection_count(address) == 0
-    connection = await pool._acquire(address, timeout=3)
+    connection = await pool._acquire(address, 3, None)
     assert pool.in_use_connection_count(address) == 1
     await pool.release(connection)
     assert pool.in_use_connection_count(address) == 0
@@ -205,10 +207,10 @@ async def test_pool_in_use_count(pool):
 async def test_pool_max_conn_pool_size(pool):
     async with AsyncFakeBoltPool((), max_connection_pool_size=1) as pool:
         address = ("127.0.0.1", 7687)
-        await pool._acquire(address, timeout=0)
+        await pool._acquire(address, 0, None)
         assert pool.in_use_connection_count(address) == 1
         with pytest.raises(ClientError):
-            await pool._acquire(address, timeout=0)
+            await pool._acquire(address, 0, None)
         assert pool.in_use_connection_count(address) == 1
 
 
@@ -222,7 +224,7 @@ async def test_pool_reset_when_released(is_reset, pool):
         with mock.patch(f"{__name__}.{quick_connection_name}.reset",
                         new_callable=AsyncMock) as reset_mock:
             is_reset_mock.return_value = is_reset
-            connection = await pool._acquire(address, timeout=3)
+            connection = await pool._acquire(address, 3, None)
             assert isinstance(connection, AsyncQuickConnection)
             assert is_reset_mock.call_count == 0
             assert reset_mock.call_count == 0